services/frame-archiver/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-pipeline-frame-archiver"
version = "0.1.0"
description = "CPU+GPU archiver — JPEG encode via CUDA IPC, detection event storage"
requires-python = ">=3.11"
dependencies = [
    "cv-pipeline-shared",
    "redis>=5.0",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.2",
]
s3 = [
    "boto3>=1.34",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "moto[s3]>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["frame_archiver"]
